backend/requirements.txt
discord.py
pytz
mysqlclient
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
django-crontab
django-q
python-dotenv
python-decouple